Factor WebSocket client into its own context
This may not be the most elegant solution in the long run, but it
solves the problem of each React component causing new websocket
connections to be opened. With this change, the server keeps track of
exactly one websocket connection for each instance of the application
that's running (i.e. 1 per tab instead of n per tab).